+++ eras/allegra/impl/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 1.7.0.0
 
+* Add `DecCBOR` instances for:
+  * `Timelock`
+  * `AllegraTxAuxData`
+  * `AllegraTxBody`
 * Converted `CertState` to a type family
 * Made the fields of predicate failures and environments lazy
 * Add `Era era` constraint to `NoThunks` instance for `TimeLock`